Implementado flex
